try to fix markdown links renaming not working, but failed (#24)
the first problem is due to wrong RegExp, which is fixed in the commit,
but the second problem that the old link cannot be replaced is because markdown links are generated like
`![](Pasted%20image%2020230128212514.png)`, the RegExp did not consider the `%20` URL encoding.

At this time I realized that the only way to solve this is to use Obsidian's API to get the link text rather than guessing it,
this is how the next commit will implement.